Add JavaFX GUI and allow user input
Refactor the main method to display a GUI created with JavaFX. This also
allows for user input without the need to use a command line. The GUI
includes a short description and screenshot showing how to get the file
path of the directory that contains files in need of renaming. In the
event of an invalid file path, no actions are taken and the program lets
you know with an error message but does not terminate.