Separate renameFile(...) into two methods
Refactor half of the code in renameFile(...) into its own method, as
previously renameFile(...) both checked if an input file had to be
renamed, as well as renamed it. Now the first half is handled by a
new method, validateFile(...) which then calls renameFile(...) when
a file's name matches a stored key word.